fix: support `swaylock` actually unlocking
Need to configure `swaylock/pam`.

Another option would be to install `sway` in the `machine` config, via
`program.sway.enable = true`, and then set `package = null` in the
home-manager sway config. I'm not currently pursuing that path, but it
may be appropriate if I find further config needs to be in machine to
get sway to work.